requirements.txt
click==8.0.4
flash==1.0.3
kornia==0.6.4
lightning_bolts==0.4.0
lightning_flash==0.7.4
matplotlib==3.4.3
nltk==3.6.6
numpy==1.21.2
omegaconf==2.1.1
Pillow==9.3.0
pydicom==2.2.2
pytorch_lightning==1.5.9
requests==2.26.0
scipy==1.7.1
segmentation_models_pytorch==0.2.1
stanza==1.3.0
timm==0.4.12
torch==1.10.1
torchvision==0.11.2
transformers==4.17.0

